# contract-swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Web.API",
    "version": "1.0"
  },
  "paths": {
    "/api/auction/create": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAuctionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAuctionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartAuctionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auction/placebid": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceBidRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceBidRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceBidRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auction/close": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseAuctionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseAuctionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloseAuctionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auction/active": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveAuctionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveAuctionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveAuctionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auction/delete": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAuctionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAuctionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAuctionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auction/{id}": {
      "get": {
        "tags": [
          "Auctions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auction": {
      "get": {
        "tags": [
          "Auctions"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicle/register": {
      "post": {
        "tags": [
          "Vehicle"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicle/update": {
      "put": {
        "tags": [
          "Vehicle"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicle": {
      "post": {
        "tags": [
          "Vehicle"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Request"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Request"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicle/{id}": {
      "get": {
        "tags": [
          "Vehicle"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicle/delete/{id}": {
      "post": {
        "tags": [
          "Vehicle"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActiveAuctionRequest": {
        "type": "object",
        "properties": {
          "auctionId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuctionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "vehicleID": {
            "type": "string",
            "nullable": true
          },
          "currentBid": {
            "type": "number",
            "format": "double"
          },
          "isActive": {
            "type": "boolean"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CloseAuctionRequest": {
        "type": "object",
        "properties": {
          "auctionId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeleteAuctionRequest": {
        "type": "object",
        "properties": {
          "auctionId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlaceBidRequest": {
        "type": "object",
        "properties": {
          "auctionId": {
            "type": "string",
            "nullable": true
          },
          "bidAmount": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "Request": {
        "type": "object",
        "properties": {
          "manufacturer": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "minPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "maxPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "year": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StartAuctionRequest": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "active": {
            "type": "boolean"
          },
          "auctionId": {
            "type": "string",
            "nullable": true
          },
          "startBid": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VehicleDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "string",
            "nullable": true
          },
          "startingBid": {
            "type": "number",
            "format": "double"
          },
          "manufacturer": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "loadCapacity": {
            "type": "string",
            "nullable": true
          },
          "numberOfSeats": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "numberOfDoors": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}